+++ src/LibHac/FsSrv/FsCreator/RomFileSystemCreator.cs
@@ -1,7 +1,7 @@
 ﻿using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
-using LibHac.FsSystem.RomFs;
+using LibHac.Tools.FsSystem.RomFs;
 
 namespace LibHac.FsSrv.FsCreator;
