+++ src/LibHac/Tools/FsSystem/NcaUtils/Nca.cs
@@ -10,8 +10,8 @@ using LibHac.Diag;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 using LibHac.FsSystem;
-using LibHac.FsSystem.RomFs;
 using LibHac.Spl;
+using LibHac.Tools.FsSystem.RomFs;
 
 namespace LibHac.Tools.FsSystem.NcaUtils;
