+++ src/LibHac/Tools/FsSystem/RomFs/RomFsFile.cs
@@ -3,7 +3,7 @@ using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
 
-namespace LibHac.FsSystem.RomFs;
+namespace LibHac.Tools.FsSystem.RomFs;
 
 public class RomFsFile : IFile
 {
