+++ src/LibHac/Tools/FsSystem/RomFs/RomFsFileSystem.cs
@@ -2,8 +2,9 @@
 using LibHac.Common;
 using LibHac.Fs;
 using LibHac.Fs.Fsa;
+using LibHac.FsSystem;
 
-namespace LibHac.FsSystem.RomFs;
+namespace LibHac.Tools.FsSystem.RomFs;
 
 public class RomFsFileSystem : IFileSystem
 {
